pyproject.toml
[tool.poetry]
name = "ai-assis"
version = "0.1.0"
description = ""
authors = ["Alan <tyujn24@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
transformers = "^4.38.1"
torch = {version = "^2.2.1+cpu", source = "cpu_torch"}
redis = "^5.0.2"
redis-dict = "^2.5.0"
timefhuman = "^0.0.5"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.2.2"


[[tool.poetry.source]]
name = "cpu_torch"
url = "https://download.pytorch.org/whl/cpu"
default = false
secondary = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
